+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - **Bug fixes**
   - Set password minimal size to 12 characters and add error feedback
+  - Display password after Reset password for temp password
 
 ## 1.6.1